+++ src/main/java/com/newbarams/ajaja/global/exception/ErrorResponse.java
@@ -1,4 +1,4 @@
-package com.newbarams.ajaja.global.common.error;
+package com.newbarams.ajaja.global.exception;
 
 import org.springframework.http.HttpStatus;
