+++ src/main/java/com/newbarams/ajaja/global/security/jwt/util/JwtParser.java
@@ -1,6 +1,6 @@
 package com.newbarams.ajaja.global.security.jwt.util;
 
-import static com.newbarams.ajaja.global.common.error.ErrorCode.*;
+import static com.newbarams.ajaja.global.exception.ErrorCode.*;
 
 import java.util.Optional;
 
@@ -9,7 +9,7 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Component;
 
-import com.newbarams.ajaja.global.common.exception.AjajaException;
+import com.newbarams.ajaja.global.exception.AjajaException;
 import com.newbarams.ajaja.global.security.common.CustomUserDetailService;
 
 import io.jsonwebtoken.Claims;
