+++ src/main/java/com/newbarams/ajaja/module/feedback/presentation/FeedbackController.java
@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.newbarams.ajaja.global.common.AjajaResponse;
-import com.newbarams.ajaja.global.common.error.ErrorResponse;
+import com.newbarams.ajaja.global.exception.ErrorResponse;
 import com.newbarams.ajaja.global.security.common.UserId;
 import com.newbarams.ajaja.module.feedback.application.GetTotalAchieveService;
 import com.newbarams.ajaja.module.feedback.application.UpdateFeedbackService;
