+++ src/main/java/com/newbarams/ajaja/module/user/application/VerifyCertificationService.java
@@ -1,6 +1,6 @@
 package com.newbarams.ajaja.module.user.application;
 
-import static com.newbarams.ajaja.global.common.error.ErrorCode.*;
+import static com.newbarams.ajaja.global.exception.ErrorCode.*;
 
 import java.util.Objects;
 
@@ -8,7 +8,7 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.newbarams.ajaja.global.common.exception.AjajaException;
+import com.newbarams.ajaja.global.exception.AjajaException;
 import com.newbarams.ajaja.module.user.application.model.EmailVerification;
 import com.newbarams.ajaja.module.user.domain.User;
