+++ src/main/java/com/newbarams/ajaja/module/user/domain/User.java
@@ -1,11 +1,11 @@
 package com.newbarams.ajaja.module.user.domain;
 
-import static com.newbarams.ajaja.global.common.error.ErrorCode.*;
+import static com.newbarams.ajaja.global.exception.ErrorCode.*;
 
 import org.hibernate.annotations.Where;
 
 import com.newbarams.ajaja.global.common.BaseEntity;
-import com.newbarams.ajaja.global.common.exception.AjajaException;
+import com.newbarams.ajaja.global.exception.AjajaException;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Embedded;
